Rename `vulkan_api.hpp` to `vk_api.hpp`
Consistant with other vulkan header patterns.
This header in particular is to configure our Vulkan API usage.
