+++ include/renderer_vk/renderer_vk.hpp
@@ -1,5 +1,5 @@
 #include "renderer.hpp"
-#include "vulkan_api.hpp"
+#include "vk_api.hpp"
 
 class GPU;
