+++ include/renderer_vk/vk_memory.hpp
@@ -5,7 +5,7 @@
 #include <utility>
 
 #include "helpers.hpp"
-#include "vulkan_api.hpp"
+#include "vk_api.hpp"
 
 namespace Vulkan {
