+++ src/core/renderer_vk/vk_api.cpp
@@ -1,3 +1,3 @@
-#include "renderer_vk/vulkan_api.hpp"
+#include "renderer_vk/vk_api.hpp"
 
 VULKAN_HPP_DEFAULT_DISPATCH_LOADER_DYNAMIC_STORAGE;
